Move laddermap things into own module. Move screenshot system to new utils module. Add untested apply_pos_fn method
